refactor: move functions to separate file
