+++ src/app/search-page/card/card.directive.ts
@@ -1,19 +1,8 @@
 import {
   Directive, ElementRef, Input, OnInit,
 } from '@angular/core';
+import { dateDifferenceInDays, dateDifferenceInMonths } from 'src/app/shared/utils/dateUtils';
 
-function dateDifferenceInDays(date1 : Date, date2: Date) {
-  const differenceInMilliseconds = date2.getTime() - date1.getTime();
-  const days = Math.ceil(differenceInMilliseconds / (1000 * 3600 * 24));
-  return days;
-}
-function dateDifferenceInMonths(date1 : Date, date2 : Date) {
-  let months = 0;
-  months = (date2.getFullYear() - date1.getFullYear()) * 12;
-  months -= date1.getMonth() + 1;
-  months += date2.getMonth();
-  return months;
-}
 
 @Directive({
   selector: '[appCard]',
